UpdateDB

+++ BatchjobService/BatchjobService/Entities/Personalizations.cs
@@ -12,7 +12,7 @@ namespace BatchjobService.Entities
         public DateTime? ModifiedDate { get; set; }
         public bool? IsSuggestion { get; set; }
         public bool? IsChosen { get; set; }
-        public int? TimeInteraction { get; set; }
+        public double? TimeInteraction { get; set; }
 
         public virtual Tags IdTagNavigation { get; set; }
         public virtual Users IdUserNavigation { get; set; }
